requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==3.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
coverage==6.5.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
importlib-metadata==5.1.0
importlib-resources==5.2.2
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==22.0
pluggy==1.2.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pycpfcnpj==1.7.0
PyJWT==2.4.0
pyrsistent==0.19.3
pytest==7.1.2
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.1.2
xmltodict==0.13.0
zipp==3.15.0
